+++ packages/web-core/src/entities/session_token.entity.ts
@@ -1,4 +1,4 @@
-const TIME_DIFF_MS = 50 * 60 * 1000; // 5 minutes
+const TIME_DIFF_MS = 5 * 60 * 1000; // 5 minutes
 
 export interface SessionTokenProps {
   token: string;