+++ packages/web-core/src/services/inworld_connection.service.ts
@@ -7,6 +7,7 @@ import {
   ConversationIntializeState,
   ConversationParticipant,
   ConversationState,
+  PerceivedLatencyReportProps,
   TriggerParameter,
 } from '../common/data_structures';
 import {
@@ -286,6 +287,12 @@ export class InworldConnectionService<
     return this.oneToOneConversation.sendNarratedAction(text);
   }
 
+  async sendPerceivedLatenctReport(props: PerceivedLatencyReportProps) {
+    await this.ensureOneToOneConversation();
+
+    return this.oneToOneConversation.sendPerceivedLatenctReport(props);
+  }
+
   async reloadScene() {
     // TODO: Remove this deprecation warning in the next major release.
     console.warn('Reload scene is deprecated. Please use changeScene instead.');
@@ -397,6 +404,10 @@ export class InworldConnectionService<
     characters?: Character[];
   }) {
     return this.connection.getEventFactory().baseProtoPacket(props);
+  }
+
+  markPacketAsHandled(packet: InworldPacketT) {
+    return this.connection.markPacketAsHandled(packet);
   }
 
   private async ensureOneToOneConversation() {